Fix the hello-world example code
Running the initial example on the demo page gives a `Children must be a list of VDOM objects or strings` which is not a great first experience with the project ;)

Python lists of VDOM objects are neither VDOM objects nor strings so they trigger the error. If we splice the results of the `fancy` call into the `h1` arguments list (so each list element becomes a separate argument to `h1`) then the code works correctly.